+++ Chapter_4/BankOOP2_ListOfAccountObjects/Main_Bank_Version2.py
@@ -33,7 +33,7 @@ accountsList[1].show()
 print()
 userName = input('What is the name for a new user account? ')
 userBalance = input('What is the starting balance for this account? ')
-userBalance = float(userBalance)
+userBalance = int(userBalance)
 userPassword = input('What is the password you want to use for this account? ')
 oAccount = Account(userName, userBalance, userPassword)
 accountsList.append(oAccount)  # append to list of accounts
